requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
openai
python-dotenv
tiktoken
langchain_chroma
